Modified build files to build in our EPICS environment. Fixed some 64 bit build warnings in server.c
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/modules/nullHttpd/current@13533 ececa60f-9f63-0410-97c7-ea05888eb260

+++ src/server.c
@@ -519,7 +519,7 @@ unsigned _stdcall htloop(void *x)
 unsigned htloop(void *x)
 #endif
 {
-	int sid=(int)x;
+	int sid=(int)(long)x;
 
 	conn[sid].id=pthread_self();
 #ifndef WIN32
@@ -602,7 +602,7 @@ void accept_loop(void *x)
 #endif
 		} else {
 			conn[i].id=1;
-			if (pthread_create(&conn[i].handle, &thr_attr, (void*(*)(void*)) htloop, (void *)i)==-1) {
+			if (pthread_create(&conn[i].handle, &thr_attr, (void * (*)(void*)) htloop, (void *)(long)i)==-1) {
 				logerror("htloop() failed...");
 				exit(0);
 			}
